:sparkles: Posts page just as a presentational component
Presentational component of the Posts page, just a few interactions are added, such as the switch between the contribution and the forum section and the count of the characters on the input. Some things have to be changed in the future: the font size at the post image component and the icons size with the distribution of the grids.